// Task_20-08-26/Decorator-based-Validation-System/frontend/src/types/user.ts
export interface User {
    id: string;
    name: string;
    email: string;
    age: number;
}

export interface CreateUserInput {
    name: string;
    email: string;
    age: number;
    password: string;
}

export interface UpdateUserInput {
    name?: string;
    email?: string;
    age?: number;
    password?: string;
}